requirements.txt
aiocoap==0.4.3
asyncio==3.4.3
click==8.1.2
Flask==2.1.1
getmac==0.8.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
paho-mqtt==1.6.1
pysqlcipher3==1.0.4
Werkzeug==2.1.1
zipp==3.8.0
